Added multithreading
