+++ data/augs.py
@@ -3,9 +3,75 @@ Augmentations and factory function for augmentations.
 
 """
 
-from models.backtranslation import BackTranslationModel
 import torch
 import numpy as np
+from typing import Dict
+import sys
+import json
+import multiprocess
+# specifies a dictionary of architectures
+_AUGS: Dict[str, any] = {}  # registry
+
+
+def register_aug(name):
+    """Decorator used register a CARP architecture
+    Args:
+        name: Name of the architecture
+    """
+
+    def register_class(cls, name):
+        _AUGS[name] = cls
+        setattr(sys.modules[__name__], name, cls)
+        return cls
+
+    if isinstance(name, str):
+        name = name.lower()
+        return lambda c: register_class(c, name)
+
+    cls = name
+    name = cls.__name__
+    register_class(cls, name.lower())
+
+    return cls
+
+
+def get_aug(name):
+    return _AUGS[name.lower()]
+
+
+def get_aug_names():
+    return _AUGS.keys()
+
+class RandomAug:
+    """
+    Randomly apply one of the augmentations.
+    """
+
+    def __init__(self, p=0.5):
+        self.p = p
+
+    def __call__(self, data, threads=16):
+        # creates a multiprocess pool to apply the augmentation
+        pool = multiprocess.Pool(threads)
+        aug_data = pool.map(self.rand_apply, data)
+        pool.close()
+        return aug_data
+
+    def rand_apply(self, datum):
+        """
+        Calls apply on the individual apply functions after first sampling torch.rand()
+        """
+        if torch.rand(1) < self.p:
+            return self.apply(datum)
+        else:
+            return datum
+    
+
+    def apply(self, data):
+        """
+        Apply augmentation to the data.
+        """
+        raise NotImplementedError
 
 
 class Compose:
@@ -13,8 +79,12 @@ class Compose:
     Compose several augmentations together. Based on torchvision's `Compose`
     """
 
-    def __init__(self, augs):
-        self.augs = augs
+    def __init__(self, composition_path):
+        #  composition_path refers to a json file that contains the augmentations we're using
+        with open(composition_path, 'r') as f:
+            composition = json.load(f)
+        for c in composition:
+            self.augs.append(get_aug(c['name'])(**c['params']))
 
     def __call__(self, data):
         for t in self.augs:
@@ -31,92 +101,7 @@ class Compose:
 
 
 
-class RandomAug:
-    """
-    Randomly apply one of the augmentations.
-    """
 
-    def __init__(self, p=0.5):
-        self.p = p
-
-    def __call__(self, data):
-        aug_data = []
-        for i in range(len(data)):
-            if torch.rand(1) < self.p:
-                aug_data.append(self.apply(data[i]))        
-            else:
-                aug_data.append(data[i])
-        return aug_data
-    
-    def apply(self, data):
-        """
-        Apply augmentation to the data.
-        """
-        raise NotImplementedError
-
-
-
-class KeyboardAug(RandomAug):
-    pass
-
-class SpellingAug(RandomAug):
-    def __init__(self, spelling_dict, p=0.5):
-        super().__init__(p)
-        self.spelling_dict = spelling_dict if type(spelling_dict) == dict else self.load_spelling_dict(spelling_dict)
-
-    def load_spelling_dict(self, file_path):
-        """
-        Loads the spelling dictionary from the file.
-        """
-        spelling_dict = {}
-        with open(file_path, 'r', encoding="utf-8") as f:
-            for line in f.readlines():
-                tokens = line.split(' ')
-                # Last token include newline separator
-                tokens[-1] = tokens[-1].replace('\n', '')
-
-                key = tokens[0]
-                values = tokens[1:]
-
-                if key not in spelling_dict:
-                    spelling_dict[key] = []
-
-                spelling_dict[key].extend(values)
-                # Remove duplicate mapping
-                spelling_dict[key] = list(set(spelling_dict[key]))
-                # Build reverse mapping
-                if self.include_reverse:
-                    for value in values:
-                        if value not in spelling_dict:
-                         spelling_dict[value] = []
-                        if key not in spelling_dict[value]:
-                         spelling_dict[value].append(key)
-        return spelling_dict
-
-    def apply(self, i):
-        """
-        Apply augmentation to the element.
-        """
-
-        # Replace the word with the correct spelling
-        if i not in self.dict:
-            return i
-        return self.dict[i]
-
-            
-
-
-
-class BackTranslationAug(RandomAug):
-    def __init__(self, src_model_name, tgt_model_name, device='cpu', batch_size=32, max_length=300, p=0.5):
-        super().__init__(p)
-        self.model = BackTranslationModel(src_model_name, tgt_model_name, device, batch_size, max_length)
-
-    def apply(self, i):
-        """
-        Apply augmentation to the element.
-        """
-        return self.model.translate(i)
